requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress bars
tqdm>=4.65.0

# Jupyter support (optional)
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0

# For Google Colab compatibility
google-colab>=1.0.0; platform_system=="Linux"
